lxd: Add more cases of VM not ready yet handling.
Recent versions of LXD will give an exit code of -1, which
translates into 255, and a different error message for the case
of LXD VM agent not being available.

LXD might reboot a guest as part of its bringup, one example being
Ubuntu 20.04 (Focal Fossa), attempts to lxc exec at this point in
time will result in a temporary "Instance is not running" error.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>